requirements.txt
# ==============================================================================
# NIDS Deep Learning & Quantum ML - Core Dependencies
# ==============================================================================
# Python 3.12 | CUDA 12.2 Compatible
# ==============================================================================

# ------------------------------------------------------------------------------
# Core Scientific Computing
# ------------------------------------------------------------------------------
numpy>=1.26.0
pandas>=2.1.0
scipy>=1.11.0
scikit-learn>=1.3.0

# ------------------------------------------------------------------------------
# Deep Learning Frameworks
# ------------------------------------------------------------------------------
# TensorFlow with GPU support
tensorflow>=2.15.0
tensorflow-datasets>=4.9.0

# PyTorch with CUDA 12.1 support (install separately with pip for CUDA)
# Use: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
torchmetrics>=1.2.0

# ------------------------------------------------------------------------------
# Data Visualization
# ------------------------------------------------------------------------------
matplotlib>=3.8.0
seaborn>=0.13.0
plotly>=5.18.0

# ------------------------------------------------------------------------------
# Jupyter & Interactive Development
# ------------------------------------------------------------------------------
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0
notebook>=7.0.0
nbformat>=5.9.0
tqdm>=4.66.0

# ------------------------------------------------------------------------------
# Network Traffic Analysis
# ------------------------------------------------------------------------------
scapy>=2.5.0
dpkt>=1.9.8

# ------------------------------------------------------------------------------
# Data Processing & Feature Engineering
# ------------------------------------------------------------------------------
imbalanced-learn>=0.11.0  # SMOTE and other resampling
category-encoders>=2.6.0
feature-engine>=1.6.0

# ------------------------------------------------------------------------------
# Model Management & Experiment Tracking
# ------------------------------------------------------------------------------
mlflow>=2.9.0
tensorboard>=2.15.0
wandb>=0.16.0

# ------------------------------------------------------------------------------
# Configuration & Utilities
# ------------------------------------------------------------------------------
pyyaml>=6.0.1
python-dotenv>=1.0.0
hydra-core>=1.3.0
omegaconf>=2.3.0
rich>=13.7.0  # Beautiful terminal output
typer>=0.9.0  # CLI framework

# ------------------------------------------------------------------------------
# Code Quality & Testing
# ------------------------------------------------------------------------------
black>=23.12.0
flake8>=6.1.0
pylint>=3.0.0
isort>=5.13.0
pytest>=7.4.0
pytest-cov>=4.1.0
mypy>=1.7.0

# ------------------------------------------------------------------------------
# File Handling
# ------------------------------------------------------------------------------
requests>=2.31.0
aiohttp>=3.9.0
h5py>=3.10.0
pyarrow>=14.0.0
openpyxl>=3.1.0
